requirement.txt
rtmidi
numpy
scipy
pyaudio
python-rtmidi
pillow